requirements.txt
asgiref
certifi
distlib
Django
django-cors-headers
django-environ
django-filter
django-ulid
djangorestframework
djangorestframework-simplejwt
filelock
markdown
pipenv
platformdirs
pyjwt
python-ulid
pytz
setuptools
sqlparse
tzdata
uuid
virtualenv
virtualenv-clone